New API

+++ src/main/java/com/leave/leaveManager/repository/LeaveTypeRepository.java
@@ -3,5 +3,5 @@ package com.leave.leaveManager.repository;
 import com.leave.leaveManager.entity.LeaveType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveTypeRepository extends JpaRepository<LeaveType, Integer>{
+public interface LeaveTypeRepository extends JpaRepository<LeaveType, Long>{
 }
